templates/boilerplate/server-py-aiohttp/requirements.txt
aiohttp
python-socketio
thrift==0.11.0